Organize files into directories
