--- styles.css
@@ -1,145 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-body {
-    font-family: Arial, sans-serif;
-    background-color: #f4f4f4;
-    color: #333;
-    line-height: 1.6;
-}
-
-header {
-    background-color: #004d99;
-    color: white;
-    padding: 1rem 0;
-    text-align: center;
-}
-
-header .logo h1 {
-    font-size: 2rem;
-    margin-bottom: 0.5rem;
-}
-
-.nav-links {
-    list-style: none;
-    display: flex;
-    justify-content: center;
-    margin: 1rem 0;
-}
-
-.nav-links li {
-    margin: 0 1.5rem;
-}
-
-.nav-links li a {
-    color: white;
-    text-decoration: none;
-    font-size: 1.1rem;
-    transition: color 0.3s ease;
-}
-
-.nav-links li a:hover {
-    color: #ffcc00;
-}
-
-.hero {
-    background-color: #0066cc;
-    color: white;
-    text-align: center;
-    padding: 4rem 2rem;
-}
-
-.hero h2 {
-    font-size: 2.5rem;
-    margin-bottom: 1rem;
-}
-
-.hero p {
-    font-size: 1.25rem;
-}
-
-.form-section {
-    background-color: white;
-    padding: 2rem;
-    max-width: 600px;
-    margin: 2rem auto;
-    border-radius: 5px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-}
-
-.form-section h3 {
-    font-size: 1.75rem;
-    text-align: center;
-    margin-bottom: 1.5rem;
-}
-
-.form-group {
-    margin-bottom: 1rem;
-}
-
-.form-group label {
-    display: block;
-    margin-bottom: 0.5rem;
-    font-weight: bold;
-}
-
-.form-group input {
-    width: 100%;
-    padding: 0.75rem;
-    font-size: 1rem;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-}
-
-.form-group input:focus {
-    border-color: #004d99;
-    outline: none;
-}
-
-.form-group button {
-    width: 100%;
-    padding: 1rem;
-    background-color: #004d99;
-    color: white;
-    border: none;
-    border-radius: 4px;
-    font-size: 1.25rem;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-}
-
-.form-group button:hover {
-    background-color: #003366;
-}
-
-footer {
-    background-color: #004d99;
-    color: white;
-    text-align: center;
-    padding: 1rem 0;
-    position: fixed;
-    width: 100%;
-    bottom: 0;
-}
-
-@media (max-width: 600px) {
-  .form-section {
-    padding: 1rem;
-    margin: 1rem;
-  }
-
-  .form-group input, .form-group button {
-    font-size: 1rem;
-  }
-
-  .hero h2 {
-    font-size: 2rem;
-  }
-
-  .hero p {
-    font-size: 1rem;
-  }
-}
